# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification - Task Management System","description":"OpenApi documentation for Task Manager","contact":{"name":"Muryginds","email":"muryginds@gmail.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Local ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/task/{id}":{"get":{"tags":["Task controller"],"summary":"Получение задания","operationId":"getTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Задание отправлено","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}},"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"400":{"description":"Некорректные параметры запроса","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}}}},"put":{"tags":["Task controller"],"summary":"Редактирование задания","operationId":"editTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequestDTO"}}},"required":true},"responses":{"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"400":{"description":"Некорректные параметры запроса","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"200":{"description":"Задание отредактировано","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}},"406":{"description":"Неверный запрос, не пройдены валидации","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}}}},"delete":{"tags":["Task controller"],"summary":"Удаление задания","operationId":"deleteTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"400":{"description":"Некорректные параметры запроса","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"200":{"description":"Задание удалено","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}}}}},"/api/v1/comment/{id}":{"get":{"tags":["Comment controller"],"summary":"Получение комментария","operationId":"getComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Комментарий отправлен","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}},"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"400":{"description":"Некорректные параметры запроса","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}}}},"put":{"tags":["Comment controller"],"summary":"Редактирование комментария","operationId":"editComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRequestDTO"}}},"required":true},"responses":{"200":{"description":"Комментарий отредактирован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}},"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"400":{"description":"Некорректные параметры запроса","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"406":{"description":"Неверный запрос, не пройдены валидации","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}}}},"post":{"tags":["Comment controller"],"summary":"Добавление комментария","operationId":"addComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRequestDTO"}}},"required":true},"responses":{"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"400":{"description":"Некорректные параметры запроса","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"200":{"description":"Комментарий добавлен","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}},"406":{"description":"Неверный запрос, не пройдены валидации","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}}}},"delete":{"tags":["Comment controller"],"summary":"Удаление комментария","operationId":"deleteComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"400":{"description":"Некорректные параметры запроса","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"200":{"description":"Комментарий удален","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}}}}},"/api/v1/task":{"post":{"tags":["Task controller"],"summary":"Добавление задания","operationId":"addTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequestDTO"}}},"required":true},"responses":{"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"400":{"description":"Некорректные параметры запроса","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"406":{"description":"Неверный запрос, не пройдены валидации","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"200":{"description":"Задание добавлено","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["Authentication"],"summary":"Регистрация пользователя","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequestDTO"}}},"required":true},"responses":{"200":{"description":"Аккаунт зарегистрирован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"406":{"description":"Неверный запрос, не пройдены валидации","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}}}}},"/api/v1/auth/authenticate":{"post":{"tags":["Authentication"],"summary":"Авторизация пользователя","operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequestDTO"}}},"required":true},"responses":{"401":{"description":"Неверный пароль или имя пользователя","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"406":{"description":"Неверный запрос, не пройдены валидации","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"200":{"description":"Пользователь авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/v1/task/{id}/recover":{"patch":{"tags":["Task controller"],"summary":"Восстановление задания","operationId":"recoverTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"400":{"description":"Некорректные параметры запроса","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"200":{"description":"Задание восстановлено","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}}}}},"/api/v1/comment/{id}/recover":{"patch":{"tags":["Comment controller"],"summary":"Восстановление комментария","operationId":"recoverComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Комментарий восстановлен","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}},"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"400":{"description":"Некорректные параметры запроса","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}}}}},"/api/v1/task/all":{"get":{"tags":["Task controller"],"summary":"Получение списка заданий","operationId":"getTasks","parameters":[{"name":"author_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"executor_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"pagesOffset","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32","default":0}},{"name":"itemPerPage","in":"query","required":false,"schema":{"minimum":1,"type":"integer","format":"int32","default":5}},{"name":"showDeleted","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"200":{"description":"Список отправлен","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}},"406":{"description":"Неверный запрос, не пройдены валидации","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}}}}},"/api/v1/comment/all":{"get":{"tags":["Comment controller"],"summary":"Получение списка комментариев","operationId":"getComments","parameters":[{"name":"task_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pagesOffset","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32","default":0}},{"name":"itemPerPage","in":"query","required":false,"schema":{"minimum":1,"type":"integer","format":"int32","default":5}},{"name":"showDeleted","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"403":{"description":"Пользователь не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}},"200":{"description":"Список отправлен","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}}},"406":{"description":"Неверный запрос, не пройдены валидации","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDTO"}}}}}}}},"components":{"schemas":{"TaskRequestDTO":{"required":["description","title"],"type":"object","properties":{"title":{"maxLength":100,"minLength":1,"type":"string","description":"Заголовок","example":"Текст нового заголовка"},"description":{"maxLength":2147483647,"minLength":1,"type":"string","description":"Описание","example":"Текст описания задания"},"status":{"type":"string","description":"Статус задания","example":"IN_PROGRESS","enum":["PENDING","IN_PROGRESS","FINISHED"]},"priority":{"type":"string","description":"Приоритет выполнения","example":"LOW","enum":["HIGHT","MEDIUM","LOW"]},"executorId":{"type":"integer","description":"id исполнителя","format":"int64","example":3}},"description":"Модель данных для создания задания"},"ErrorResponseDTO":{"type":"object","properties":{"errors":{"type":"array","description":"Список ошибок","example":"Ошибка входных данных","items":{"type":"string","description":"Список ошибок","example":"Ошибка входных данных"}}},"description":"Модель данных отправки ошибок"},"CommentDTO":{"type":"object","properties":{"id":{"type":"integer","description":"id комментария","format":"int64"},"text":{"type":"string","description":"Текст"},"author":{"$ref":"#/components/schemas/UserDTO"},"deleted":{"type":"boolean"}},"description":"Модель данных комментария"},"UserDTO":{"type":"object","properties":{"id":{"type":"integer","description":"id пользователя","format":"int64","example":2},"name":{"type":"string","description":"Имя","example":"Дмитрий"},"email":{"type":"string","description":"Почта","example":"muryginds@gmail.com"},"token":{"type":"string","description":"Токен авторизации"}},"description":"Модель данных пользователя"},"CommentRequestDTO":{"required":["text"],"type":"object","properties":{"text":{"maxLength":500,"minLength":1,"type":"string","description":"Текст комментария","example":"Текст нового комментария"}},"description":"Модель данных для создания комментария"},"RegisterRequestDTO":{"required":["name","password"],"type":"object","properties":{"name":{"maxLength":200,"minLength":1,"type":"string","description":"Имя пользователя","example":"Дмитрий"},"password":{"maxLength":50,"minLength":1,"type":"string","description":"Пароль","example":"secretpassword"},"email":{"pattern":"^[a-zA-Z0-9_!#$%&'*+/=?`{|}~^.-]+@[a-zA-Z0-9.-]+$","type":"string","description":"Email пользователя","example":"muryginds@gmail.com"}},"description":"Модель данных для регистрации нового пользователя"},"AuthRequestDTO":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","description":"Пользовательский email","example":"muryginds@gmail.com"},"password":{"type":"string","description":"Пользовательский пароль"}},"description":"Модель данных для авторизации"}},"securitySchemes":{"bearerAuth":{"type":"http","description":"authentication using JWT token","scheme":"bearer","bearerFormat":"JWT"}}}}